extract set_public_address
