+++ validator/src/main.rs
@@ -221,41 +221,7 @@ pub fn main() {
             return;
         }
         ("set-public-address", Some(subcommand_matches)) => {
-            let parse_arg_addr = |arg_name: &str, arg_long: &str| -> Option<SocketAddr> {
-                subcommand_matches.value_of(arg_name).map(|host_port| {
-                    solana_net_utils::parse_host_port(host_port).unwrap_or_else(|err| {
-                        eprintln!(
-                            "Failed to parse --{arg_long} address. It must be in the HOST:PORT \
-                             format. {err}"
-                        );
-                        exit(1);
-                    })
-                })
-            };
-            let tpu_addr = parse_arg_addr("tpu_addr", "tpu");
-            let tpu_forwards_addr = parse_arg_addr("tpu_forwards_addr", "tpu-forwards");
-
-            macro_rules! set_public_address {
-                ($public_addr:expr, $set_public_address:ident, $request:literal) => {
-                    if let Some(public_addr) = $public_addr {
-                        let admin_client = admin_rpc_service::connect(&ledger_path);
-                        admin_rpc_service::runtime()
-                            .block_on(async move {
-                                admin_client.await?.$set_public_address(public_addr).await
-                            })
-                            .unwrap_or_else(|err| {
-                                eprintln!("{} request failed: {err}", $request);
-                                exit(1);
-                            });
-                    }
-                };
-            }
-            set_public_address!(tpu_addr, set_public_tpu_address, "setPublicTpuAddress");
-            set_public_address!(
-                tpu_forwards_addr,
-                set_public_tpu_forwards_address,
-                "setPublicTpuForwardsAddress"
-            );
+            commands::set_public_address::execute(subcommand_matches, &ledger_path);
             return;
         }
         _ => unreachable!(),
